CLUE-603: take the reviewable items from the second review
Teale's inline comments, all of them mechanical.

The resolver's `any` cast is gone, and removing it disproved a claim this branch
had put in the docs. TileContentModel itself calls tileContentAPIViews({}) -- "add
empty apis so they are available on the generic type" -- and every content model
extends it, so getObjectsForVariable is always present. The doc said roughly half
the content models don't route through that and the optional call was therefore
load-bearing; the opposite is true. A test whose stub modelled a tile that "does
not implement the hook at all" was modelling a state that cannot occur, and now
models the real one.

IHighlightTarget is no longer a second declaration of the same shape. It stays as
a name -- a resolved highlight target and an annotation endpoint are the same
shape for different reasons -- but it is now an alias of IClueObjectSnapshot, so
the two cannot drift.

The variable resolver moves to shared-variables-registration, with its tests, so
core no longer knows what a variable is. Only the plugin registers that kind now,
which means a document with no variables plugin loaded resolves variable
references to nothing rather than resolving them by accident.

objectHighlightState and isObjectHighlighted take optional ids, so callers stop
repeating the same guard.
