+++ src/models/highlights/highlight-reference.ts
@@ -1,14 +1,14 @@
+import type { IClueObjectSnapshot } from "../annotations/clue-object";
 // Resolvers only need to walk the document's tiles, so they take the base content model.
 import type { BaseDocumentContentModelType } from "../document/base-document-content";
 
 /**
- * A resolved, concrete thing a tile can render emphasis on.
+ * A resolved, concrete thing a tile can render emphasis on. Structurally identical to the
+ * annotation system's object address, and aliased rather than redeclared so the two cannot drift
+ * apart silently — but named for what it is here, because a resolved highlight target and an
+ * annotation endpoint are the same shape for different reasons.
  */
-export interface IHighlightTarget {
-  tileId: string;
-  objectId: string;
-  objectType?: string;
-}
+export type IHighlightTarget = IClueObjectSnapshot;
 
 /**
  * A reference to something that should be highlighted. Deliberately NOT named
@@ -61,19 +61,4 @@ export function sameHighlightReference(a: HighlightReference, b: HighlightRefere
 registerReferenceResolver("object", ref => {
   if (ref.kind !== "object") return [];
   return [{ tileId: ref.tileId, objectId: ref.objectId, objectType: ref.objectType }];
-});
-
-registerReferenceResolver("variable", (ref, content) => {
-  if (ref.kind !== "variable") return [];
-  const targets: IHighlightTarget[] = [];
-  content.tileMap.forEach(tile => {
-    // Tiles opt in by implementing getObjectsForVariable; the rest are skipped. The cast is
-    // needed because tile.content is the union of every registered tile content model.
-    const tileContent = tile.content as any;
-    const objects = tileContent?.getObjectsForVariable?.(ref.variableId);
-    objects?.forEach((object: { objectId: string; objectType?: string }) => {
-      targets.push({ tileId: tile.id, objectId: object.objectId, objectType: object.objectType });
-    });
-  });
-  return targets;
 });
